新增中间表价目表信息同步

+++ yunya-models/src/main/java/com/yunya/models/middletable/BaseTariffInfo.java
@@ -1,7 +1,9 @@
 package com.yunya.models.middletable;
 
+import javax.persistence.Column;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import java.math.BigDecimal;
-import javax.persistence.*;
 
 @Table(name = "base_tariff_info")
 public class BaseTariffInfo {
@@ -17,7 +19,7 @@ public class BaseTariffInfo {
      */
     @Id
     @Column(name = "item_type")
-    private Byte itemType;
+    private Integer itemType;
 
     /**
      * 父项目分类
@@ -76,7 +78,7 @@ public class BaseTariffInfo {
      *
      * @return item_type - 项目类型（0-价目表；1-商品）
      */
-    public Byte getItemType() {
+    public Integer getItemType() {
         return itemType;
     }
 
@@ -85,7 +87,7 @@ public class BaseTariffInfo {
      *
      * @param itemType 项目类型（0-价目表；1-商品）
      */
-    public void setItemType(Byte itemType) {
+    public void setItemType(Integer itemType) {
         this.itemType = itemType;
     }
